pyproject.toml
[project]
name = "storytellers"
version = "0.1.0"
description = "Add your description here"
authors = [{ name = "Ben Swift", email = "ben@benswift.me" }]
dependencies = [
    "diffusers>=0.30.2",
    "transformers>=4.44.2",
    "accelerate>=0.34.2",
    "pillow>=8.0.0",
    "opencv-python-headless>=4.5.0",
]
readme = "README.md"
requires-python = ">= 3.12"

[project.scripts]
"storytellers" = "storytellers:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/storytellers"]
